+++ stroom-core-client/src/main/java/stroom/data/client/presenter/DataPresenter.java
@@ -602,6 +602,9 @@ public class DataPresenter
             updateAvailableAndSelectedTabs(streamTypeName, availableChildStreamTypes);
             refreshMetaInfoPresenterContent(currentSourceLocation.getMetaId());
             refreshProgressBar(false);
+            // As we are not hitting the rest service we need to clear this out else it will
+            // think we are on a strm/part that we are not.
+            lastResult = null;
         } else {
             // Tabs will be updated by updateFromResource
             setEffectiveChildStreamType(streamTypeName, availableChildStreamTypes);
@@ -682,7 +685,6 @@ public class DataPresenter
 
     private boolean isSameStreamAndPartAsLastTime() {
         if (lastResult != null) {
-
             final Long lastId = Optional.ofNullable(lastResult)
                     .flatMap(result -> Optional.ofNullable(result.getSourceLocation()))
                     .map(SourceLocation::getMetaId)
@@ -692,8 +694,13 @@ public class DataPresenter
                     .map(SourceLocation::getPartIndex)
                     .orElse(null);
 
-            return Objects.equals(getCurrentMetaId(), lastId)
+            final boolean isSame = Objects.equals(getCurrentMetaId(), lastId)
                     && Objects.equals(getCurrentPartIndex(), lastPartNo);
+//            GWT.log(lastId + ":" + lastPartNo
+//                    + " => "
+//                    + getCurrentMetaId() + ":" + getCurrentPartIndex()
+//                    + " = " + isSame);
+            return isSame;
         } else {
             return false;
         }