+++ stroom-core-client/src/main/java/stroom/explorer/client/presenter/ExplorerDropDownTreePresenter.java
@@ -69,7 +69,7 @@ class ExplorerDropDownTreePresenter
                         view.setQuickFilterTooltipSupplier(() -> QuickFilterTooltipUtil.createTooltip(
                                 "Choose Item Quick Filter",
                                 ExplorerTreeFilter.FIELD_DEFINITIONS,
-                                uiConfig.getHelpUrl())));
+                                uiConfig.getHelpUrlQuickFilter())));
     }
 
     public void show() {