+++ stroom-legacy/stroom-legacy-impex_6_1/src/main/java/stroom/legacy/impex_6_1/MappingUtil.java
@@ -676,6 +676,7 @@ public final class MappingUtil {
         return new stroom.processor.shared.QueryData(
                 map(value.getDataSource()),
                 map(value.getExpression()),
+                null,
                 map(value.getLimits()));
     }