+++ stroom-lmdb/src/main/java/stroom/lmdb/serde/Deserializer.java
@@ -15,7 +15,7 @@
  *
  */
 
-package stroom.lmdb;
+package stroom.lmdb.serde;
 
 import java.nio.ByteBuffer;