+++ stroom-pipeline/src/main/java/stroom/pipeline/refdata/store/offheapstore/serdes/RangeStoreKeySerde.java
@@ -18,7 +18,7 @@
 package stroom.pipeline.refdata.store.offheapstore.serdes;
 
 import stroom.bytebuffer.ByteBufferUtils;
-import stroom.lmdb.Serde;
+import stroom.lmdb.serde.Serde;
 import stroom.pipeline.refdata.store.offheapstore.RangeStoreKey;
 import stroom.pipeline.refdata.store.offheapstore.UID;
 import stroom.util.logging.LambdaLogger;