+++ stroom-query/stroom-query-common/src/main/java/stroom/query/common/v2/EventSearch.java
@@ -1,18 +1,19 @@
 package stroom.query.common.v2;
 
 import stroom.query.api.v2.Query;
+import stroom.task.api.TaskContext;
 
 import java.util.concurrent.CompletableFuture;
-import java.util.function.Consumer;
+import java.util.function.BiConsumer;
 
 public interface EventSearch {
 
-    CompletableFuture<EventRefs> search(Query query,
-                                        EventRef minEvent,
-                                        EventRef maxEvent,
-                                        long maxStreams,
-                                        long maxEvents,
-                                        long maxEventsPerStream,
-                                        int resultSendFrequency,
-                                        Consumer<EventRefs> consumer);
+    CompletableFuture<Void> search(TaskContext parentTaskContext,
+                                   Query query,
+                                   EventRef minEvent,
+                                   EventRef maxEvent,
+                                   long maxStreams,
+                                   long maxEvents,
+                                   long maxEventsPerStream,
+                                   BiConsumer<EventRefs, Throwable> consumer);
 }