+++ stroom-search/stroom-search-elastic/src/main/java/stroom/search/elastic/search/ElasticSearchConfig.java
@@ -13,22 +13,12 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 @JsonPropertyOrder(alphabetic = true)
 public class ElasticSearchConfig extends AbstractConfig implements IsStroomConfig {
 
-    /**
-     * We don't want to collect more than 10k doc's data into the queue by
-     * default. When the queue is full the index shard data tasks will pause
-     * until the docs are drained from the queue.
-     */
-    private static final int DEFAULT_MAX_STORED_DATA_QUEUE_SIZE = 1000;
-    private static final int DEFAULT_MAX_BOOLEAN_CLAUSE_COUNT = 1024;
-
-    private final int maxStoredDataQueueSize;
-    private final int maxBooleanClauseCount;
+    private final StroomDuration scrollDuration;
     private final String storeSize;
     private final CacheConfig searchResultCache;
 
     public ElasticSearchConfig() {
-        maxStoredDataQueueSize = DEFAULT_MAX_STORED_DATA_QUEUE_SIZE;
-        maxBooleanClauseCount = DEFAULT_MAX_BOOLEAN_CLAUSE_COUNT;
+        scrollDuration = StroomDuration.ofMinutes(1);
         storeSize = "1000000,100,10,1";
         searchResultCache = CacheConfig.builder()
                 .maximumSize(10000L)
@@ -38,25 +28,17 @@ public class ElasticSearchConfig extends AbstractConfig implements IsStroomConfi
 
     @SuppressWarnings("unused")
     @JsonCreator
-    public ElasticSearchConfig(@JsonProperty("maxStoredDataQueueSize") final int maxStoredDataQueueSize,
-                               @JsonProperty("maxBooleanClauseCount") final int maxBooleanClauseCount,
+    public ElasticSearchConfig(@JsonProperty("scrollDuration") final StroomDuration scrollDuration,
                                @JsonProperty("storeSize") final String storeSize,
                                @JsonProperty("searchResultCache") final CacheConfig searchResultCache) {
-        this.maxStoredDataQueueSize = maxStoredDataQueueSize;
-        this.maxBooleanClauseCount = maxBooleanClauseCount;
+        this.scrollDuration = scrollDuration;
         this.storeSize = storeSize;
         this.searchResultCache = searchResultCache;
     }
 
-    @JsonPropertyDescription("The maximum number documents that will have stored data retrieved from the index " +
-            "shard and queued prior to further processing")
-    public int getMaxStoredDataQueueSize() {
-        return maxStoredDataQueueSize;
-    }
-
-    @JsonPropertyDescription("The maximum number of clauses that a boolean search can contain.")
-    public int getMaxBooleanClauseCount() {
-        return maxBooleanClauseCount;
+    @JsonPropertyDescription("Amount of time to allow an Elasticsearch scroll request to continue before aborting.")
+    public StroomDuration getScrollDuration() {
+        return scrollDuration;
     }
 
     @JsonPropertyDescription("The maximum number of search results to keep in memory at each level.")
@@ -71,9 +53,8 @@ public class ElasticSearchConfig extends AbstractConfig implements IsStroomConfi
     @Override
     public String toString() {
         return "ElasticSearchConfig{" +
-                "maxStoredDataQueueSize=" + maxStoredDataQueueSize +
-                ", maxBooleanClauseCount=" + maxBooleanClauseCount +
-                ", storeSize='" + storeSize + '\'' +
+                "scrollDuration='" + scrollDuration + "'" +
+                ", storeSize=" + storeSize +
                 ", searchResultCache=" + searchResultCache +
                 '}';
     }